Changing the loader to assimp from miniEngine.

+++ Plugins/src/mini/scene/plugins/fbx/mesh/FBXLayerElement.java
@@ -67,7 +67,7 @@ public class FBXLayerElement {
 
     public static FBXLayerElement fromElement(FBXElement fbxElement) {
         FBXLayerElement layerElement = new FBXLayerElement();
-        layerElement.index = (int) (long) fbxElement.getProperties().get(0);
+        layerElement.index = (int) fbxElement.getProperties().get(0);
 
         String elementType = fbxElement.name.substring("LayerElement".length());
         layerElement.type = Type.valueOf(elementType);
